+++ internal/loadertest/fakeloader.go
