+++ pkg/transformers/config/factorycrd_test.go
@@ -20,9 +20,9 @@ import (
 	"reflect"
 	"testing"
 
+	"sigs.k8s.io/kustomize/internal/loadertest"
 	"sigs.k8s.io/kustomize/pkg/gvk"
 	"sigs.k8s.io/kustomize/pkg/ifc"
-	"sigs.k8s.io/kustomize/pkg/internal/loadertest"
 )
 
 // This defines two CRD's:  Bee and MyKind.
